requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
beautifulsoup4
selenium
webdriver-manager
pyautogui
openpyxl 